fix(build): separate client parser from server journal-service and fix build type errors
